requirements.txt
python-dotenv==1.0.0
langchain-openai
langchain-core
pinecone-client==2.2.1
